Se fixeo el test de nivel ya que instanciaba asientos con objetos en null. Esto traía problema con la implementación de hash y equals de la misma clase.
Se creó una exception en el adaptador de lanchita para que en lugar de devolver una lista vacía, tire una exception.

+++ aterrizar.com/src/ar/com/aterrizar/modelo/Asiento.java
@@ -54,7 +54,7 @@ public class Asiento {
 	public int hashCode() {
 		int enteroDisponible= 0;
 		if(this.disponible) enteroDisponible = 1;
-		return (int)this.tipo + (int)this.ubicacion + (int)this.aerolinea.hashCode() + this.precio.hashCode() + this.codigo.hashCode()+ enteroDisponible ;
+		return (int)this.tipo + (int)this.ubicacion + this.aerolinea.hashCode() + this.precio.hashCode() + this.codigo.hashCode()+ enteroDisponible ;
 	}
 	
 	@Override
